+++ tests/site.rs
@@ -118,6 +118,10 @@ fn can_build_site_without_live_reload() {
     assert!(file_exists!(public, "posts/tutorials/programming/index.html"));
     // TODO: add assertion for syntax highlighting
 
+    // aliases work
+    assert!(file_exists!(public, "an-old-url/old-page/index.html"));
+    assert!(file_contains!(public, "an-old-url/old-page/index.html", "something-else"));
+
     // No tags or categories
     assert_eq!(file_exists!(public, "categories/index.html"), false);
     assert_eq!(file_exists!(public, "tags/index.html"), false);