+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Add `section` to a page Tera context if there is one
 - Reverse `order` sorting to be more intuitive: they are now desc, think of them
 as 1st, 2nd in the list
+- Add `aliases` to pages for when you are changing urls but want to redirect
+to the new one
 
 ## 0.0.6 (2017-05-24)